Make Dimon link against libmri.
This is a critical piece that I cannot check (besides the fact that it
compiles and starts) since it provides realtime functionality for a
scanner and environment that I don't have access to. Dimon is compiled
"without libmri" for this purpose, which of course contradicts the aim
of using a shared library.

The original code would cause problems due to duplicate symbols from
mri_dicom_hdr.c and associated headers.